+++ CHANGELOG.md
@@ -107,3 +107,7 @@ Initial release of `slim-tools` with the following features:
 - Added the signature help provider (`signature-help.ts`) to provide hints on parameters in functions/methods
 - Added the code actions provider (`code-actions.ts`) to provide quick fixes for common errors and simple refactorings
 - Improved the validation service to avoid throwing errors while a line is being typed
+
+## [0.0.15] (CAT)
+- Added the rename provider (`rename.ts`) to manage refactoring in large documents/codebases
+- Added the formatting provider (`formatting.ts`) for formatting of documents and code chunks